Refactor to use deployList function
